Fix empty apps thumbnails on 8.1
to manage the new hardware bitmaps in the right way, we now
use the ImageView canvas to set the resized app screenshot instead
of creating a new canvas.
Doing this let's also do the bitmap resizing in the ImageView onDraw
(so when we really show the image in the recents panel) and not as soon
as we load the hardware screenshot for all recents task, with a nice
benefit for panel loading  performances.

For now let's skip also the thumbnail cache and see how the panel behaves
without it.

Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>