Fallback to main activity icon if iconpack doesn't provide any specific activity icon
See for example youtube icon in the recents panel when using Elun icon pack
when youtube is in a specific video window.
It's not the main Youtube activity icon from the icon pack. That's because icon packs
have specific icons for specific app activities, so if you are in an activity of the app
that is not provided by the icon pack, the recents panel will use the default system icon for it.

So for those cases, we can fallback to the main activity icon provided by the iconpack
for the package.
This won't cover all cases, but only for packages that have a "MainActivity" in the component name.
Better than nothing.

Change-Id: I0fad9ddd8187abf92a361e44fb97bc46f91ed645
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>